+++ src/libfcimodels.h
@@ -14,12 +14,6 @@ void Ji_J_J7(const double q[7], double b_Ji_J_J7[42]);
 void Ji_J_J8(const double q[7], double b_Ji_J_J8[42]);
 void Ji_J_J9(const double q[7], const double F_T_EE[16], double b_Ji_J_J9[42]);
 
-void M_NE(const double q[7],
-          const double I_load[9],
-          double m_load,
-          const double F_x_Cload[3],
-          double M_NE[49]);
-
 void O_J_J1(double b_O_J_J1[42]);
 void O_J_J2(const double q[7], double b_O_J_J2[42]);
 void O_J_J3(const double q[7], double b_O_J_J3[42]);
@@ -39,17 +33,4 @@ void O_T_J6(const double q[7], double b_O_T_J6[16]);
 void O_T_J7(const double q[7], double b_O_T_J7[16]);
 void O_T_J8(const double q[7], double b_O_T_J8[16]);
 void O_T_J9(const double q[7], const double F_T_EE[16], double b_O_T_J9[16]);
-
-void c_NE(const double q[7],
-          const double dq[7],
-          const double I_load[9],
-          double m_load,
-          const double F_x_Cload[3],
-          double c_NE[7]);
-
-void g_NE(const double q[7],
-          const double g_earth[3],
-          double m_load,
-          const double F_x_Cload[3],
-          double g_NE[7]);
 }